pyproject.toml
[project]
name = "devtree"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx",
    "pandas>=2.3.3",
    "pillow",
    "pydantic",
    "pymupdf>=1.26.7",
    "pytesseract",
    "rapidfuzz>=3.14.3",
    "tqdm",
]
